pybuild: do not invoke clean commands for missing interpreters
clean commands are invoked for requested (f.e. via Build-Depends)
interpreters only, but if given interpreter is not installed, it usually
also means there are no files to clean - it should be safe to skip
these commands.

It's a common practice to build source package outside build chroot
so this change will make it easier to build pybuild based packages
(host machines will not have to install these interpreters)