version: Fix a problem in handling ranges with no minimum
When presented with a version range with no minimum,
get_requested_versions would try to set the minimum to 0.0, which would
fail because of a condition in the Version constructor that erroneously
confuses major=0 with the major version being unset. This change fixes
the problem and adds a couple of doctests to demonstrate the issue and
prevent regressions.